Everything is working! Ready to turn in. I should get meets expectations. This commit includes database populator. From 3 json files you can populate data and make changes to the database easily.
in the Fixes after submit.txt you will find the latest changes.

Things to do:
Now we can add created and updated functions when commiting changes to db.
With this we should be able to implement api endpoints exceeds. Look into
it.

This and csrf tokens should get me to exceeds expectations.